refactor: move Left and Right to index file
